ENH:Millora de la sincronització de llesques, ja que es té en compte la distància recorreguda envers del nombre de llesques incrementat. S'ha deshabilitat la sincronització del window level i el zoom perquè els metges no volen que estigui per defecte.

+++ starviewer/src/core/synchronizetooldata.cpp
@@ -21,13 +21,13 @@ SynchronizeToolData::~SynchronizeToolData()
 {
 }
 
-void SynchronizeToolData::setIncrement( int value )
+void SynchronizeToolData::setIncrement( double value )
 {
     m_increment = value;
     emit( sliceChanged() );
 }
 
-int SynchronizeToolData::getIncrement()
+double SynchronizeToolData::getIncrement()
 {
     return m_increment;
 }
